Added source code folder for math4610
Currently only houses 3 functions, a main, a TaylorSeriesExpansion, and a printToFile function. The print function is not working properly in the gradle format. It was printing to the screen twice (once for SOUT statements, and once for the file print statements) I have put the results into a file for now, but if the printToFile function gets fixed it will be updated.